C: Clarify source install section
The intent of the source installation section was to demonstrate how a
source install _may_ look like. Because this seems to lead to confusion,
I've instead simply indicated to read the bundled README in the source
archive which contains the proper information, and is always up-to-date
with the given archive.