Added subtitle information to the grouping for right hand nav.
In the future, the ingested file will render a pulldown menu for
the different grouped things. They also need a way to show a
subtitle indicating what versions.

The solution here is to give the ingest something to split on, a
";" in this case, and then it can extract and render whatever
string is there as a subheading if it exists.