Remove old 'sdks' content fixing crossreferences.
Leaving old DITA content around is dangerous as it can let
the build complete fine, but the end result file is either
included or there will be a broken link.

Since the sdks/ content should no longer be used, I removed
it and fixed the references to the right places.