fixing some of the wording...
I'll let you reword this, but I wanted to clarify the wording so you knew what I meant by the comment in the pull request.  User could, conceivably, still use old SDK with 5.0 but it will be a workaround (as you note in your legacy section) and not best practice.  We want to tell them they _must_ upgrade SDK when upgrading to 5.0.  They can upgrade SDK and still use it with 4.6, no problem though.