Add more specifics around the Request Data expected
It may not be clear that the Request Data should be in the body of the POST as a form parameter. Someone passing the values in the URL or as JSON will get a possibly confusing response.

Since this is common in most of the endpoint documentation, do you think it would be a good idea to create a new documentation page under the REST area to explain the details of making a HTTP POST/PUT/GET etc and just link to that instead of having to put these details on every REST doc page?